backend/pyproject.toml
[tool.poetry]
name = "freelancequest"
version = "0.1.0"
description = "My bachelor's final project"
authors = ["codelotusr <smaliukasdeividas@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
channels = "^4.2.2"
channels-redis = "^4.2.1"
django-cors-headers = "^4.7.0"
django-environ = "^0.12.0"
redis = "^5.2.1"
psycopg2 = "^2.9.10"
daphne = "^4.1.2"
django-filter = "^25.1"
drf-nested-routers = "^0.94.1"
watchgod = "^0.8.2"
whitenoise = "^6.9.0"
drf-spectacular = "^0.28.0"
djangorestframework-simplejwt = {extras = ["django"], version = "^5.5.0"}
django-allauth = "^65.7.0"
dj-rest-auth = "^7.0.1"
pillow = "^11.2.1"
requests = "^2.32.3"
django-stubs = "^5.1.3"
django-stubs-ext = "^5.1.3"
django-extensions = "^4.1"
graphviz = "^0.20.3"
pydotplus = "^2.0.2"
pylint = "^3.3.7"
pytest-cov = "^6.1.1"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.1.0"
django-stubs = "^5.1.3"
django-stubs-ext = "^5.1.3"
django-extensions = "^4.0"
pytest = "^8.3.5"
pytest-django = "^4.11.1"
factory-boy = "^3.3.3"
faker = "^37.1.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "freelancequest.settings"
python_files = ["test_*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
